+++ src/receiver.rs
@@ -1,15 +1,15 @@
-//! Measurement receiver that actually processes all measurements coming from services.
+//! Readings receiver that actually processes all readings coming from services.
 
 use crate::db::*;
-use crate::measurement::*;
+use crate::reading::*;
 use log::info;
 use std::sync::mpsc::Receiver;
 use std::sync::{Arc, Mutex};
 
 /// Run the receiver.
-pub fn run(rx: Receiver<Measurement>, db: Arc<Mutex<Db>>) {
-    for measurement in rx {
-        info!("{}: {:?}", &measurement.sensor, &measurement.value);
-        db.lock().unwrap().insert_measurement(&measurement);
+pub fn run(rx: Receiver<Reading>, db: Arc<Mutex<Db>>) {
+    for reading in rx {
+        info!("{}: {:?}", &reading.sensor, &reading.value);
+        db.lock().unwrap().insert_reading(&reading);
     }
 }