+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New: generic persisted JSON key-value store
 - Break: add `expires_ts` to the key-value store
+- Break: rename `Measurement` into `Reading`
 
 ## `0.7.0`